pyproject.toml
[tool.poetry]
name = "descidb"
version = "0.1.0"
description = "A database management system for document processing and embedding"
authors = ["Vardhan Shorewala <vardhanshorewala@coophive.com>"]
readme = "README.md"
packages = [{include = "descidb"}]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["database", "document-processing", "embedding"]
maintainers = ["Vardhan Shorewala <vardhanshorewala@coophive.com>"]

[tool.poetry.dependencies]
python = ">=3.9.0,<4.0"
neo4j = "5.27.0"
certifi = "2025.1.31"
requests = "2.32.3"
psycopg2 = "2.9.9"
numpy = ">=1.22.5,<2.0.0"
chromadb = "0.5.3"
python-dotenv = "1.0.1"
openai = "1.63.0"
PyPDF2 = "3.0.1"
transformers = "4.48.3"
web3 = "7.2.0"
flask = "3.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.0"
pytest-cov = "4.1.0"
pytest-mock = "3.11.1"
pytest-env = "0.8.2"
black = "23.7.0"
isort = "5.12.0"
mypy = "1.5.1"
flake8 = "6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
descidb = "descidb.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=descidb --cov-report=xml --cov-report=term"

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.coverage.run]
source = ["descidb"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
] 